Grasping at Threads
Working on threading the game starting with the loading at the beginning to get it up and running faster.  Still running into errors with the skybox.

GameObjects are now just a holder for a Model and its instances. 
Player no longer uses a separate shader from other game objects.

ResourceManager now keeps copies of all textures, materials and models for quick access later.